Fixed package naming issue with example app
