+++ example/build.gradle
@@ -9,7 +9,7 @@ android {
     buildToolsVersion "23.0.2"
 
     defaultConfig {
-        applicationId "io.github.btkelly.gandalf"
+        applicationId "io.github.btkelly.gandalf.example"
         minSdkVersion 16
         targetSdkVersion 23
         versionCode 1
@@ -18,6 +18,7 @@ android {
 
     lintOptions {
         abortOnError true
+        warning 'InvalidPackage'
     }
 
     testOptions.unitTests {
@@ -45,9 +46,12 @@ license {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
+
+    compile fileTree(include: ['*.jar'], dir: 'libs')
     compile 'com.android.support:appcompat-v7:23.1.1'
+    compile project(':gandalf')
+
+    testCompile 'junit:junit:4.12'
 }
 
 preBuild.dependsOn licenseFormat
