+++ example/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="io.github.btkelly.gandalf">
+    package="io.github.btkelly.gandalf.example">
 
     <application
         android:allowBackup="true"
