+++ example/src/main/java/io/github/btkelly/gandalf/example/DeleteMe.java
@@ -13,21 +13,10 @@
  * License for the specific language governing permissions and limitations
  * under the License.
  */
-package io.github.btkelly.gandalf;
-
-import org.junit.Assert;
-import org.junit.Test;
-
+package io.github.btkelly.gandalf.example;
 
 /**
- * To work on unit tests, switch the Test Artifact in the Build Variants view.
+ * TODO: Add a class header comment!
  */
-public class ExampleUnitTest {
-
-    private static final int EXPECTED = 4;
-
-    @Test
-    public void additionIsCorrect() {
-        Assert.assertEquals(EXPECTED, 2 + 2);
-    }
+public class DeleteMe {
 }
